Renaming to address #265
